Move zornable to layer-agnostic utilities
