+++ packages/voictents-and-estinants-engine/src/layer-agnostic-utilities/deprecated-zorn/getZornableId.ts
@@ -1,5 +1,5 @@
 import { getTextDigest } from '../../package-agnostic-utilities/string/getTextDigest';
-import { Zornable } from '../../adapted-programs/programmable-units/in-memory-cache/zornable';
+import { Zornable } from './zornable';
 
 // eslint-disable-next-line @typescript-eslint/no-use-before-define
 export const getZornableId = (zornable: Zornable): string => {
