+++ packages/voictents-and-estinants-engine/src/layer-agnostic-utilities/deprecated-zorn/zornable.ts
@@ -1,8 +1,8 @@
 import { Merge } from 'type-fest';
-import { Voque } from '../../../core/engine/voque';
-import { AbstractSerializableIndexByName } from '../../../core-programs/abstractSerializableVoictent';
-import { Gepp } from '../../../core/engine-shell/voictent/gepp';
-import { Hubblepup } from '../../../core/engine-shell/hubblepup/hubblepup';
+import { Voque } from '../../core/engine/voque';
+import { AbstractSerializableIndexByName } from '../../core-programs/abstractSerializableVoictent';
+import { Gepp } from '../../core/engine-shell/voictent/gepp';
+import { Hubblepup } from '../../core/engine-shell/hubblepup/hubblepup';
 
 export type Zornable = {
   zorn: string;
